file restructure and update images with lower size
