+++ src/pages/landingPages/donate/components/companyProgressSection/companyProgress.scss
@@ -3,71 +3,72 @@
 }
 
 #progress-container {
-    max-width: var(--content-width);
-    margin: 5rem auto;
+	max-width: var(--content-width);
+	margin: 10rem auto;
 
-    .progress-image-wrapper {
-        position: relative;
-        
-        .progress-image {
-            margin: 0 auto;
-            padding: 1rem;
-        }
+	.progress-image-wrapper {
+		position: relative;
+		margin-top: 2rem;
 
-        .progress-card {
-            position: absolute;   
-            
-            @media (max-width: 1400px) {
-                transform: scale(0.8);
-            }
+		.progress-image {
+			margin: 0 auto;
+			padding: 1rem;
+		}
 
-            .progress-percentage-container{
-                color: var(--text-color);
-                display: flex;
-                margin-bottom: 0;
-                font-weight: 500;
+		.progress-card {
+			position: absolute;
 
-                .progress-percentage{               
-                    align-self: flex-end;
-                }
-            }
+			@media (max-width: 1400px) {
+				transform: scale(0.8);
+			}
 
-            p {
-                text-align: left;
-            }
+			.progress-percentage-container {
+				color: var(--text-color);
+				display: flex;
+				margin-bottom: 0;
+				font-weight: 500;
 
-            .percentage-up-arrow {
-                max-height: 44px;
-                max-width: 44px;
-            }
-        }
+				.progress-percentage {
+					align-self: flex-end;
+				}
+			}
 
-        .progress-card-1 {
-            top: 32%;
-            right: 7%;
+			p {
+				text-align: left;
+			}
 
-            @media (min-width: 1400px) {
-                right: 8%;
-            }
-        }
+			.percentage-up-arrow {
+				max-height: 44px;
+				max-width: 44px;
+			}
+		}
 
-        .progress-card-2 {
-            top: 12%;
-            left: 8%;
+		.progress-card-1 {
+			top: 32%;
+			right: 7%;
 
-            @media (min-width: 1400px) {
-                left: 14%;
-            }
-        }
+			@media (min-width: 1400px) {
+				right: 8%;
+			}
+		}
 
-        .progress-card-3 {
-            white-space: pre-line;
-            bottom: 2%;
-            left: 46%;
+		.progress-card-2 {
+			top: 12%;
+			left: 8%;
 
-            @media (min-width: 1400px) {
-                bottom: 5%;
-            }
-        }
-    }   
+			@media (min-width: 1400px) {
+				left: 14%;
+			}
+		}
+
+		.progress-card-3 {
+			white-space: pre-line;
+			bottom: 2%;
+			left: 46%;
+
+			@media (min-width: 1400px) {
+				bottom: 5%;
+			}
+		}
+	}
 }
