+++ src/pages/landingPages/donate/components/donatePaymentForm/donatePaymentForm.scss
@@ -5,11 +5,12 @@
 	--secondary-text-color: #6b7888;
 	--large-btn-radius: 30px;
 	--spacing: 2rem;
+	--form-width: 970px;
 }
 
 #donate-payment-wrapper {
 	width: 100%;
-	background-image: url("../../assets/spacelab-donate-background.png");
+	background-image: url("../../assets/SpaceLab_DonatePage1_Background.png");
 	background-size: cover;
 	background-position: center;
 	background-repeat: no-repeat;
@@ -18,7 +19,7 @@
 }
 
 .donate-payment-form {
-	max-width: 970px;
+	max-width: var(--form-width);
 	padding: 2rem 0;
 	margin: 0 auto;
 }
@@ -40,91 +41,7 @@
 	}
 }
 
-.preset-amount-wrapper {
-	display: flex;
-	justify-content: stretch;
-	align-items: center;
-	flex-wrap: wrap;
-}
-
-.radio-btn-input {
-	position: absolute;
-	clip: rect(0px, 0px, 0px, 0px);
-	pointer-events: none;
-	width: 1px;
-	height: 1px;
-}
-
-.radio-input-label {
-	font-weight: bold;
-	color: var(--primary-text-color);
-	background-color: var(--secondary-background-color);
-	max-width: 30%;
-	min-width: 156px;
-	min-height: 3em;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	text-align: center;
-	margin-top: 0.5rem;
-	margin-right: 1.5rem;
-	margin-bottom: 0;
-	position: relative;
-	border-radius: 15px;
-	border: none;
-
-	@media (min-width: 600px) {
-		&:last-child {
-			margin-right: -1.5rem;
-		}
-	}
-}
-
-input:checked + .radio-input-label {
-	color: white;
-	background-color: var(--background-color);
-}
-
-.radio-input-span {
-	color: var(--secondary-text-color);
-	font-size: 21px;
-	line-height: 24px;
-	font-weight: 100;
-	margin-top: 0.5rem;
-}
-
-input:checked + .radio-input-label .radio-input-span {
-	color: white;
-}
-
-//checkBox
-.donate-checkbox-wrapper {
-	margin: 1rem 0;
-	position: relative;
-	text-align: start;
-
-	.donate-checkbox {
-		width: 1.5rem;
-		height: 1.5rem;
-		margin: 0;
-		position: absolute;
-		border: 2px solid #d9d9d9;
-	}
-
-	.cover-fee {
-		display: inline-block;
-		padding-inline-start: 2.375rem;
-		position: relative;
-		text-align: start;
-		margin: 0;
-
-		.cover-fee-bold {
-			font-weight: bold;
-		}
-	}
-}
-
-//big buttons
+//buttons
 .btn-container {
 	margin-top: 2rem;
 }
@@ -157,49 +74,6 @@ input:checked + .radio-input-label .radio-input-span {
 	&:hover,
 	&:focus {
 		background-color: transparent;
-		outline: none;
-	}
-}
-
-// toggle group buttons
-.tab-group-container {
-	width: 100%;
-}
-
-.tab-group {
-	display: flex;
-	align-items: stretch;
-	justify-content: flex-start;
-	flex-wrap: wrap;
-	border-radius: var(--large-btn-radius);
-	background-color: var(--secondary-background-color);
-}
-
-.text-caption-tab {
-	flex-grow: 1;
-	border-radius: inherit;
-	color: var(--secondary-text-color);
-	background-color: var(--secondary-background-color);
-	border: none;
-	transition: background-color 1s ease-out;
-	text-transform: none;
-	padding: 1.25rem 3rem;
-
-	&:hover {
-		background-color: var(--secondary-background-color);
-	}
-
-	&:focus {
-		outline: none;
-	}
-}
-
-.active {
-	background-color: #5700c6;
-	color: #e2e2e2;
-
-	&:hover {
-		background-color: #5700c6;
 	}
 }
 
@@ -212,88 +86,6 @@ input:checked + .radio-input-label .radio-input-span {
 	}
 }
 
-/* Accordion Btn*/
-.accordion-btn {
-	width: 100%;
-	border-radius: var(--large-btn-radius);
-	justify-content: flex-start;
-	font-family: "Lato", sans-serif;
-
-	input {
-		border-bottom: none;
-		padding: 0;
-		margin: 0;
-
-		&:disabled {
-			color: var(--secondary-text-color);
-		}
-	}
-
-	.multiline-input {
-		width: 100%;
-		height: 5rem;
-		border: none;
-		background-color: transparent;
-		outline: none;
-		color: var(--secondary-text-color);
-	}
-}
-
-.add_icon {
-	width: 1.5rem;
-	height: 1.5rem;
-	margin-right: 1rem;
-}
-
-// card payment info
-.form-container {
-	text-align: left;
-
-	input {
-		border: 2px solid var(--secondary-background-color);
-		color: var(--secondary-text-color);
-		border-radius: 10px;
-		padding: 1.25rem 3rem;
-		background-color: var(--secondary-background-color);
-	}
-
-	.form-content {
-		display: flex;
-		flex-direction: column;
-		gap: var(--spacing);
-
-		.form-input {
-			border: 2px solid var(--secondary-background-color);
-			color: var(--secondary-text-color);
-			border-radius: 10px;
-			padding: 1.25rem 3rem;
-			background-color: var(--secondary-background-color);
-		}
-
-		.input-error {
-			border-color: red;
-		}
-	}
-
-	.small-input {
-		width: 100%;
-		display: inline-flex;
-		flex-direction: row;
-		justify-content: space-between;
-		gap: 1rem;
-	}
-
-	.margin-top {
-		margin-top: var(--spacing);
-	}
-}
-
-.form-sub-container {
-	.margin-top {
-		margin-top: calc(1.5 * var(--spacing));
-	}
-}
-
 // logo placement
 .icon-container {
 	img {
@@ -302,7 +94,7 @@ input:checked + .radio-input-label .radio-input-span {
 	}
 	display: none;
 
-	@media (min-width: 1600px) {
+	@media (min-width: 896px) {
 		display: flex;
 		align-self: flex-start;
 		transform: translate(-22rem, -25rem);
@@ -314,7 +106,7 @@ input:checked + .radio-input-label .radio-input-span {
 	position: absolute;
 	display: none;
 
-	@media (min-width: 1600px) {
+	@media (min-width: 896px) {
 		display: block;
 	}
 }
