+++ src/pages/landingPages/donate/components/radioButtonGroup/radioButtonGroup.scss
@@ -1,0 +1,56 @@
+.preset-amount-wrapper {
+	display: flex;
+	justify-content: stretch;
+	align-items: center;
+	flex-wrap: wrap;
+}
+
+.radio-btn-input {
+	position: absolute;
+	clip: rect(0px, 0px, 0px, 0px);
+	pointer-events: none;
+	width: 1px;
+	height: 1px;
+}
+
+.radio-input-label {
+	font-weight: bold;
+	color: var(--primary-text-color);
+	background-color: var(--secondary-background-color);
+	max-width: 30%;
+	min-width: 156px;
+	min-height: 3em;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	text-align: center;
+	margin-top: 0.5rem;
+	margin-right: 1.5rem;
+	margin-bottom: 0;
+	position: relative;
+	border-radius: 15px;
+	border: none;
+
+	@media (min-width: 600px) {
+		&:last-child {
+			margin-right: -1.5rem;
+		}
+	}
+}
+
+input:checked + .radio-input-label {
+	color: white;
+	background-color: var(--background-color);
+}
+
+.radio-input-span {
+	color: var(--secondary-text-color);
+	font-size: 21px;
+	line-height: 24px;
+	font-weight: 100;
+	margin-top: 0.5rem;
+}
+
+input:checked + .radio-input-label .radio-input-span {
+	color: white;
+}
