+++ src/pages/landingPages/donate/components/toggleGroupButton/toggleGroupButton.scss
@@ -1,0 +1,40 @@
+.tab-group-container {
+	width: 100%;
+}
+
+.tab-group {
+	display: flex;
+	align-items: stretch;
+	justify-content: flex-start;
+	flex-wrap: wrap;
+	border-radius: var(--large-btn-radius);
+	background-color: var(--secondary-background-color);
+}
+
+.text-caption-tab {
+	flex-grow: 1;
+	border-radius: inherit;
+	color: var(--secondary-text-color);
+	background-color: var(--secondary-background-color);
+	border: none;
+	transition: background-color 1s ease-out;
+	text-transform: none;
+	padding: 1.25rem 3rem;
+
+	&:hover {
+		background-color: var(--secondary-background-color);
+	}
+
+	&:focus {
+		outline: none;
+	}
+}
+
+.active {
+	background-color: #5700c6;
+	color: #e2e2e2;
+
+	&:hover {
+		background-color: #5700c6;
+	}
+}
