+++ src/pages/landingPages/donate/donate.scss
@@ -61,7 +61,7 @@
 
 .section-background {
 	width: 100%;
-	background-image: url("./assets/spacelab-donate-background.png");
+	background-image: url("./assets/SpaceLab_DonatePage1_Background.png");
 	background-size: cover;
 	background-position: center;
 	background-repeat: no-repeat;
